Add unit tests and rework almost everything
Achieve 100% unit test coverage because it would be hypocritical to
write a testing tool that didn't test itself. `setup.py test` or `tox`
will test against python 2 and 3 in a virtualenv. This required a
refactor of basically everything to add explicit interfaces and allow
dependencies to be injected.

The CLI interface is consolidated into just a `cinderblock` command,
with subcommands `cinderblock trigger` and `cinderblock update`. This
should make it easier to extend the program without a proliferation of
commands in the future. Resolves T559.